Tile Map Service background image function added.
git-svn-id: http://my.redmine.jp/svn/iric/v3.X/gui/trunk@6121 2cc7cdd0-1db9-4218-aa4a-1d8dad43e0f0

+++ libs/post/post2d/datamodel/post2dwindowstreamlineunstructuredsettingdialog.cpp
@@ -230,7 +230,7 @@ void Post2dWindowStreamlineUnstructuredSettingDialog::removeData()
 	if (item != nullptr) {delete item;}
 	ui->startPositionListWidget->blockSignals(false);
 	m_unstSettings.erase(m_unstSettings.begin() + current);
-	if (current >= m_unstSettings.size()) { current = m_unstSettings.size() - 1; }
+	if (current >= m_unstSettings.size()) {current = static_cast<int>(m_unstSettings.size()) - 1;}
 	ui->startPositionListWidget->setCurrentRow(current);
 	m_activeSetting = &(m_unstSettings[current]);
 	applySettings();
